Created launch auth page to have combined register and login
