+++ frontend/app/contact/page.tsx
@@ -67,7 +67,6 @@ export default function ContactPage() {
     setValue,
     watch,
   } = useForm<ContactFormInputs>({
-    // @ts-expect-error – @hookform/resolvers v5 type mismatch with zod v4.3.x
     resolver: zodResolver(contactSchema),
     mode: "onBlur",
     defaultValues: {
