+++ frontend/app/login/page.tsx
@@ -1,44 +1,19 @@
 "use client";
 
-import React, { useState } from "react";
+import React from "react";
 import { useRouter } from "next/navigation";
 import { useToast } from "@/context/ToastContext";
 import { useWallet } from "@/context/WalletContext";
 import { useWizard } from "@/context/WizardContext";
 import { useAuth } from "@/app/context/AuthContext";
+import LoginForm from "@/components/auth/LoginForm";
 
 export default function LoginPage() {
   const router = useRouter();
   const { addToast } = useToast();
   const { disconnect } = useWallet();
   const { reset } = useWizard();
-  const { login, logout } = useAuth();
-
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [showPassword, setShowPassword] = useState(false);
-  const [rememberMe, setRememberMe] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const [emailError, setEmailError] = useState("");
-  const [passwordError, setPasswordError] = useState("");
-
-  const validateEmail = () => {
-    if (!email) {
-      setEmailError("Email is required");
-    } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email)) {
-      setEmailError("Please enter a valid email address");
-    } else {
-      setEmailError("");
-    }
-  };
-
-  const validatePassword = () => {
-    if (!password) {
-      setPasswordError("Password is required");
-    } else {
-      setPasswordError("");
-    }
-  };
+  const { logout } = useAuth();
 
   const handleLogout = () => {
     logout();
@@ -52,34 +27,6 @@ export default function LoginPage() {
     (window as Window & { __stellarproof_logout?: () => void }).__stellarproof_logout = handleLogout;
   }
 
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    validateEmail();
-    validatePassword();
-    if (!email || !/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email) || !password) return;
-    setIsLoading(true);
-    try {
-      await login(email, password);
-      addToast({ type: "success", message: "Welcome back!" });
-      router.push("/dashboard");
-    } catch {
-      addToast({
-        type: "error",
-        message: "Authentication failed. Please check your credentials.",
-      });
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const emailInputClass = emailError
-    ? "w-full px-4 py-2.5 rounded-lg border border-red-500 text-sm bg-white dark:bg-darkblue-dark text-gray-900 dark:text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-red-500 transition"
-    : "w-full px-4 py-2.5 rounded-lg border border-gray-300 dark:border-gray-600 text-sm bg-white dark:bg-darkblue-dark text-gray-900 dark:text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-primary transition";
-
-  const passwordInputClass = passwordError
-    ? "w-full px-4 py-2.5 rounded-lg border border-red-500 text-sm bg-white dark:bg-darkblue-dark text-gray-900 dark:text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-red-500 transition pr-10"
-    : "w-full px-4 py-2.5 rounded-lg border border-gray-300 dark:border-gray-600 text-sm bg-white dark:bg-darkblue-dark text-gray-900 dark:text-white placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-primary transition pr-10";
-
   return (
     <main id="main-content" className="min-h-screen flex items-center justify-center bg-darkblue dark:bg-darkblue-dark px-4">
       <div className="w-full max-w-md bg-white dark:bg-darkblue rounded-2xl shadow-glow p-8">
@@ -89,118 +36,9 @@ export default function LoginPage() {
             Sign in to your account
           </p>
         </div>
-        <form onSubmit={handleSubmit} noValidate className="space-y-5">
-          <div>
-            <label
-              htmlFor="email"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-            >
-              Email Address
-            </label>
-            <input
-              id="email"
-              type="email"
-              autoComplete="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              onBlur={validateEmail}
-              placeholder="you@example.com"
-              aria-invalid={!!emailError}
-              aria-describedby={emailError ? "email-error" : undefined}
-              className={emailInputClass}
-            />
-            {emailError && (
-              <p id="email-error" role="alert" className="text-red-500 text-xs mt-1">{emailError}</p>
-            )}
-          </div>
-          <div>
-            <label
-              htmlFor="password"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-            >
-              Password
-            </label>
-            <div className="relative">
-              <input
-                id="password"
-                type={showPassword ? "text" : "password"}
-                autoComplete="current-password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                onBlur={validatePassword}
-                placeholder="••••••••"
-                aria-invalid={!!passwordError}
-                aria-describedby={passwordError ? "password-error" : undefined}
-                className={passwordInputClass}
-              />
-              <button
-                type="button"
-                onClick={() => setShowPassword((prev) => !prev)}
-                aria-label={showPassword ? "Hide password" : "Show password"}
-                aria-controls="password"
-                className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-400 hover:text-gray-600 dark:hover:text-gray-200 text-xs focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary rounded"
-              >
-                <span aria-hidden="true">{showPassword ? "Hide" : "Show"}</span>
-              </button>
-            </div>
-            {passwordError && (
-              <p id="password-error" role="alert" className="text-red-500 text-xs mt-1">{passwordError}</p>
-            )}
-          </div>
-          <div className="flex items-center justify-between">
-            <label className="flex items-center gap-2 text-sm text-gray-600 dark:text-gray-400 cursor-pointer">
-              <input
-                type="checkbox"
-                checked={rememberMe}
-                onChange={(e) => setRememberMe(e.target.checked)}
-                className="accent-primary"
-              />
-              Remember me
-            </label>
-            <a
-              href="/forgot-password"
-              className="text-sm text-primary hover:underline"
-            >
-              Forgot your password?
-            </a>
-          </div>
-          <button
-            type="submit"
-            disabled={isLoading}
-            aria-label={isLoading ? "Signing in, please wait" : "Sign in"}
-            aria-busy={isLoading}
-            className="w-full py-2.5 rounded-lg bg-primary text-white text-sm font-semibold hover:bg-primary-dark transition-all duration-200 flex items-center justify-center gap-2 disabled:opacity-60 shadow-button-glow focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:ring-offset-2"
-          >
-            {isLoading ? (
-              <>
-                <svg
-                  className="animate-spin h-4 w-4 text-white"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  aria-hidden="true"
-                >
-                  <circle
-                    className="opacity-25"
-                    cx="12"
-                    cy="12"
-                    r="10"
-                    stroke="currentColor"
-                    strokeWidth="4"
-                  />
-                  <path
-                    className="opacity-75"
-                    fill="currentColor"
-                    d="M4 12a8 8 0 018-8v8z"
-                  />
-                </svg>
-                <span aria-hidden="true">Signing in...</span>
-              </>
-            ) : (
-              "Sign In"
-            )}
-          </button>
-        </form>
+        
+        <LoginForm />
+
         <p className="text-center text-sm text-gray-500 dark:text-gray-400 mt-6">
           Don&apos;t have an account?{" "}
           <a href="/register" className="text-primary font-medium hover:underline">
