+++ frontend/app/register/page.tsx
@@ -1,75 +1,9 @@
 "use client";
 
-import React, { useState } from "react";
-import { useRouter } from "next/navigation";
-import { useForm } from "react-hook-form";
-import { useToast } from "@/context/ToastContext";
-import { useAuth } from "@/app/context/AuthContext";
-
-interface RegisterFormValues {
-  fullName: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-}
-
-function getPasswordStrength(password: string) {
-  let score = 0;
-  if (password.length >= 8) score++;
-  if (password.length >= 12) score++;
-  if (/[A-Z]/.test(password)) score++;
-  if (/[0-9]/.test(password)) score++;
-  if (/[^A-Za-z0-9]/.test(password)) score++;
-
-  if (score <= 1) return { label: "Weak", color: "bg-red-500", width: "w-1/4" };
-  if (score === 2)
-    return { label: "Fair", color: "bg-yellow-500", width: "w-2/4" };
-  if (score === 3)
-    return { label: "Strong", color: "bg-blue-500", width: "w-3/4" };
-  return { label: "Very Strong", color: "bg-green-500", width: "w-full" };
-}
+import React from "react";
+import RegisterForm from "@/components/auth/RegisterForm";
 
 export default function RegisterPage() {
-  const router = useRouter();
-  const { addToast } = useToast();
-  const { register: registerAuth } = useAuth();
-  const [showPassword, setShowPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const {
-    register,
-    handleSubmit,
-    watch,
-    formState: { errors },
-  } = useForm<RegisterFormValues>({
-    mode: "onBlur",
-  });
-
-  const passwordValue = watch("password", "");
-  const strength = passwordValue ? getPasswordStrength(passwordValue) : null;
-
-  const onSubmit = async (data: RegisterFormValues) => {
-    setIsLoading(true);
-    try {
-      await registerAuth(data.fullName, data.email, data.password);
-      addToast({ type: "success", message: "Registration successful" });
-      router.push("/dashboard");
-    } catch {
-      addToast({
-        type: "error",
-        message: "Something went wrong. Please try again.",
-      });
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const inputBase =
-    "w-full px-4 py-2.5 rounded-lg border text-sm bg-white dark:bg-darkblue-dark text-gray-900 dark:text-white placeholder-gray-400 focus:outline-none focus:ring-2 transition";
-  const inputNormal = `${inputBase} border-gray-300 dark:border-gray-600 focus:ring-primary`;
-  const inputError = `${inputBase} border-red-500 focus:ring-red-500`;
-
   return (
     <div className="min-h-screen flex items-center justify-center bg-darkblue dark:bg-darkblue-dark px-4 py-12">
       <div className="w-full max-w-md bg-white dark:bg-darkblue rounded-2xl shadow-glow p-8">
@@ -80,212 +14,7 @@ export default function RegisterPage() {
           </p>
         </div>
 
-        <form
-          onSubmit={handleSubmit(onSubmit)}
-          noValidate
-          className="space-y-5"
-        >
-          {/* Full Name */}
-          <div>
-            <label
-              htmlFor="fullName"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-            >
-              Full Name
-            </label>
-            <input
-              id="fullName"
-              type="text"
-              autoComplete="name"
-              placeholder="John Doe"
-              className={errors.fullName ? inputError : inputNormal}
-              {...register("fullName", {
-                required: "Full name is required",
-                minLength: {
-                  value: 2,
-                  message: "Full name must be at least 2 characters",
-                },
-              })}
-            />
-            {errors.fullName && (
-              <p className="text-red-500 text-xs mt-1">
-                {errors.fullName.message}
-              </p>
-            )}
-          </div>
-
-          {/* Email */}
-          <div>
-            <label
-              htmlFor="email"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-            >
-              Email Address
-            </label>
-            <input
-              id="email"
-              type="email"
-              autoComplete="email"
-              placeholder="you@example.com"
-              className={errors.email ? inputError : inputNormal}
-              {...register("email", {
-                required: "Email is required",
-                pattern: {
-                  value: /^[^\s@]+@[^\s@]+\.[^\s@]+$/,
-                  message: "Please enter a valid email address",
-                },
-              })}
-            />
-            {errors.email && (
-              <p className="text-red-500 text-xs mt-1">
-                {errors.email.message}
-              </p>
-            )}
-          </div>
-
-          {/* Password */}
-          <div>
-            <label
-              htmlFor="password"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-            >
-              Password
-            </label>
-            <div className="relative">
-              <input
-                id="password"
-                type={showPassword ? "text" : "password"}
-                autoComplete="new-password"
-                placeholder="••••••••"
-                className={`${errors.password ? inputError : inputNormal} pr-14`}
-                {...register("password", {
-                  required: "Password is required",
-                  minLength: {
-                    value: 8,
-                    message: "Password must be at least 8 characters",
-                  },
-                  validate: {
-                    hasUppercase: (v) =>
-                      /[A-Z]/.test(v) ||
-                      "Password must contain at least one uppercase letter",
-                    hasNumber: (v) =>
-                      /[0-9]/.test(v) ||
-                      "Password must contain at least one number",
-                    hasSpecial: (v) =>
-                      /[^A-Za-z0-9]/.test(v) ||
-                      "Password must contain at least one special character",
-                  },
-                })}
-              />
-              <button
-                type="button"
-                onClick={() => setShowPassword((prev) => !prev)}
-                className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-400 hover:text-gray-600 dark:hover:text-gray-200 text-xs"
-              >
-                {showPassword ? "Hide" : "Show"}
-              </button>
-            </div>
-            {errors.password && (
-              <p className="text-red-500 text-xs mt-1">
-                {errors.password.message}
-              </p>
-            )}
-            {/* Password Strength Indicator */}
-            {strength && (
-              <div className="mt-2">
-                <div className="h-1.5 w-full bg-gray-200 dark:bg-gray-700 rounded-full overflow-hidden">
-                  <div
-                    className={`h-full ${strength.color} ${strength.width} rounded-full transition-all duration-300`}
-                  />
-                </div>
-                <p
-                  className={`text-xs mt-1 ${
-                    strength.label === "Weak"
-                      ? "text-red-500"
-                      : strength.label === "Fair"
-                        ? "text-yellow-500"
-                        : strength.label === "Strong"
-                          ? "text-blue-500"
-                          : "text-green-500"
-                  }`}
-                >
-                  {strength.label}
-                </p>
-              </div>
-            )}
-          </div>
-
-          {/* Confirm Password */}
-          <div>
-            <label
-              htmlFor="confirmPassword"
-              className="block text-sm font-medium text-gray-700 dark:text-gray-300 mb-1"
-            >
-              Confirm Password
-            </label>
-            <div className="relative">
-              <input
-                id="confirmPassword"
-                type={showConfirmPassword ? "text" : "password"}
-                autoComplete="new-password"
-                placeholder="••••••••"
-                className={`${errors.confirmPassword ? inputError : inputNormal} pr-14`}
-                {...register("confirmPassword", {
-                  required: "Please confirm your password",
-                  validate: (v) =>
-                    v === passwordValue || "Passwords do not match",
-                })}
-              />
-              <button
-                type="button"
-                onClick={() => setShowConfirmPassword((prev) => !prev)}
-                className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-400 hover:text-gray-600 dark:hover:text-gray-200 text-xs"
-              >
-                {showConfirmPassword ? "Hide" : "Show"}
-              </button>
-            </div>
-            {errors.confirmPassword && (
-              <p className="text-red-500 text-xs mt-1">
-                {errors.confirmPassword.message}
-              </p>
-            )}
-          </div>
-
-          {/* Submit Button */}
-          <button
-            type="submit"
-            disabled={isLoading}
-            className="w-full py-2.5 rounded-lg bg-primary text-white text-sm font-semibold hover:bg-primary-dark transition-all duration-200 flex items-center justify-center gap-2 disabled:opacity-60 shadow-button-glow"
-          >
-            {isLoading ? (
-              <>
-                <svg
-                  className="animate-spin h-4 w-4 text-white"
-                  xmlns="http://www.w3.org/2000/svg"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                >
-                  <circle
-                    className="opacity-25"
-                    cx="12"
-                    cy="12"
-                    r="10"
-                    stroke="currentColor"
-                    strokeWidth="4"
-                  />
-                  <path
-                    className="opacity-75"
-                    fill="currentColor"
-                    d="M4 12a8 8 0 018-8v8z"
-                  />
-                </svg>
-                Creating account...
-              </>
-            ) : (
-              "Create Account"
-            )}
-          </button>
-        </form>
+        <RegisterForm />
 
         <p className="text-center text-sm text-gray-500 dark:text-gray-400 mt-6">
           Already have an account?{" "}
