+++ frontend/components/CallToAction.tsx
@@ -58,10 +58,10 @@ export default function CallToAction() {
             
         
             <Link 
-              href="/register" 
+              href="/launch" 
               className="w-full sm:w-auto flex items-center justify-center gap-2 px-8 py-3.5 rounded-full bg-primary text-white font-semibold hover:bg-blue-600 transition-colors shadow-button-glow"
             >
-              Get Started
+              Launch App
               <ArrowRight className="w-5 h-5" />
             </Link>
